flask_app_v2/requirements.txt
# Requirements for flask_app_v2
Flask>=2.0.1
Flask-SQLAlchemy>=3.0.0
SQLAlchemy>=2.0.0
psycopg2-binary>=2.9.5
Flask-Migrate>=4.0.0
authlib>=1.0.0
Flask-Bcrypt>=1.0.1
Flask-Bcrypt>=1.0.1
Flask-Login>=0.6.0
celery>=5.3.0
redis>=4.5.0
python-dotenv>=1.0.0
boto3>=1.26.0
requests>=2.28.0
gradio_client>=0.2.0
python-pptx>=0.6.21
Werkzeug>=2.0.0
gunicorn>=20.1.0
pytest>=7.0.0
flask-wtf>=1.0.0
email_validator>=2.0.0
